Remove welcome modal, move Latest Imagery button
